requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
google-api-python-client==2.149.0
google-auth==2.35.0

python-dotenv==1.0.1
python-dateutil==2.9.0.post0

langchain==0.1.13
langchain-core==0.1.52
langchain-google-genai==0.0.8
google-generativeai==0.3.2
langchain-community==0.0.38

pytz==2024.2
